Refactor Embedding: deleted facade package giving its responsibility to service class, and deleted client package intead now module has dependency of OpenAI module's clietWrapper class #5
