+++ Embedding/src/main/kotlin/com/tastecompass/embedding/service/EmbeddingServiceImpl.kt
@@ -1,0 +1,38 @@
+package com.tastecompass.embedding.service
+
+import com.tastecompass.domain.entity.Restaurant
+import com.tastecompass.embedding.dto.EmbeddingResult
+import com.tastecompass.embedding.mapper.EmbeddingRequestMapper
+import com.tastecompass.openai.client.OpenAIClientWrapper
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Service
+
+@Service
+class EmbeddingServiceImpl(
+    private val client: OpenAIClientWrapper
+): EmbeddingService {
+    override suspend fun embed(
+        restaurant: Restaurant
+    ): EmbeddingResult = coroutineScope {
+        val request = EmbeddingRequestMapper.fromRestaurant(restaurant)
+
+        val moodVectorDeferred = async { client.embed(request.mood) }
+        val tasteVectorDeferred = async { client.embed(request.taste) }
+
+        logger.info("Embedding... ${restaurant.toString()}")
+
+        val moodVector = moodVectorDeferred.await()
+        val tasteVector = tasteVectorDeferred.await()
+
+        EmbeddingResult(
+            moodVector = moodVector,
+            tasteVector = tasteVector
+        )
+    }
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(this::class.simpleName)
+    }
+}
